Fix #811: Error:Failed to create pg_restore list file : out of memory
Instead of loading the contents of the list file before parsing and
transforming, this change set uses the stream parsing method by introducing
`FileIterator` and `ArchiveIterator` concepts.
